requirements.txt
aiofiles==0.8.0
aiohttp==3.8.4
aiomultiprocess==0.9.0
aiomysql==0.1.1
aiosignal==1.3.1
aiosqlite==0.17.0
async-generator==1.10
async-lru==2.0.2
async-timeout==4.0.2
asyncio-extras==1.3.2
asyncpraw==7.7.0
asyncprawcore==2.3.0
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
llvmlite==0.40.0
multidict==6.0.4
nest-asyncio==1.5.6
numpy==1.24.3
PyMySQL==1.0.3
python-dotenv==1.0.0
requests==2.31.0
typing_extensions==4.6.2
update-checker==0.18.0
urllib3==2.0.2
yarl==1.9.2
